refactor: colocate interpreter tests with plugins, final cleanup
Move num, str, and eq interpreter tests from tests/interpreters/ into
their respective tests/plugins/{name}/ directories. Only core.test.ts
remains in tests/interpreters/ since it tests the shared core interpreter.
