+++ tests/plugins/str/interpreter.test.ts
@@ -1,8 +1,8 @@
 import { describe, expect, it } from "vitest";
-import { composeInterpreters, ilo } from "../../src/core";
-import { coreInterpreter } from "../../src/interpreters/core";
-import { str } from "../../src/plugins/str";
-import { strInterpreter } from "../../src/plugins/str/interpreter";
+import { composeInterpreters, ilo } from "../../../src/core";
+import { coreInterpreter } from "../../../src/interpreters/core";
+import { str } from "../../../src/plugins/str";
+import { strInterpreter } from "../../../src/plugins/str/interpreter";
 
 function injectInput(node: any, input: Record<string, unknown>): any {
   if (node === null || node === undefined || typeof node !== "object") return node;
